refactor(structure): split database module into 3 separate responbilities
refactor(structure): データベースモジュールを3つの実務に分割

- Extract connection management into connection.py
- Extract PostgreChatMessageHistory initialization into history_manager.py
- Extract conversation title queries into conversation_queries.py
- 接続管理を connection.py に分割しました
- PostgreChatMessageHistory の初期化を history_manager.py に分離しました
- 会話タイトルのクエリ処理を conversation_queries.py に分離しました
